Moving salesman.py to examples folder.
